+++ clients/admin-ui/src/types/api/models/HealthCheck.ts
@@ -2,6 +2,7 @@
 /* tslint:disable */
 /* eslint-disable */
 
+import type { DictionaryStatus } from "./DictionaryStatus";
 import type { SystemScannerStatus } from "./SystemScannerStatus";
 
 /**
@@ -12,4 +13,5 @@ export type HealthCheck = {
   fidesplus_version: string;
   fidesplus_server: string;
   system_scanner: SystemScannerStatus;
+  dictionary: DictionaryStatus;
 };